fix: multiple routes in "list issues" section
