+++ lib/endpoint/find-description.js
@@ -1,0 +1,12 @@
+module.exports = findDescription
+
+function findDescription (state) {
+  const description = state.blocks
+    .filter(block => block.type === 'description')
+    .map(block => block.text)
+    .join('\n\n')
+
+  state.results.forEach(result => {
+    result.description = description
+  })
+}
