+++ lib/endpoint/html-to-json.js
@@ -3,11 +3,12 @@ module.exports = htmlToJson
 const readFileSync = require('fs').readFileSync
 const joinPath = require('path').join
 
-const _ = require('lodash')
 const cheerio = require('cheerio')
-const markdownTable = require('markdown-table')
 
 const elementToBlock = require('./element-to-block')
+const findDescription = require('./find-description')
+const findParameters = require('./find-parameters')
+const findRoutes = require('./find-routes')
 const htmlContainsEndpoint = require('./html-contains-endpoint')
 
 function htmlToJson (html) {
@@ -50,7 +51,7 @@ function htmlToJson (html) {
   // remove title block so we can keep track of blocks we haven’t parsed yet
   state.blocks.splice(0, 1)
 
-  // mutate result
+  // mutate results
   findRoutes(state)
   findParameters(state)
   findDescription(state)
@@ -65,187 +66,4 @@ function findOverride (html) {
 
   const overridePath = html.match(/custom override at (.*)/).pop()
   return JSON.parse(readFileSync(joinPath(process.cwd(), overridePath), 'utf8'))
-}
-
-/**
- *
- */
-function findRoutes (state) {
-  const result = state.results[0]
-  const routeBlocks = state.blocks.filter(block => block.type === 'route')
-
-  // TODO: Find descriptions and set params for multi-route sections
-  if (routeBlocks.length === 1) {
-    const routeBlock = routeBlocks[0]
-    const {method, path} = routeBlock
-    result.method = method
-    result.path = path
-    state.blocks.splice(state.blocks.indexOf(routeBlock), 1)
-    return
-  }
-
-  const hasAuthenticatedUserRoute = !!routeBlocks.slice(1).find(block => {
-    const prevBlock = state.blocks[state.blocks.indexOf(block) - 1]
-    return prevBlock.type === 'description' && /^\/user\//.test(block.path)
-  })
-  const hasExampleRoutes = !!routeBlocks.slice(1).find(block => {
-    const prevBlock = state.blocks[state.blocks.indexOf(block) - 1]
-    return prevBlock.type === 'description' && /\bexample\b/i.test(prevBlock.text)
-  })
-  const hasStubbedRoute = !!routeBlocks.slice(1).find(block => {
-    const prevBlock = state.blocks[state.blocks.indexOf(block) - 1]
-
-    return /\bstubbed endpoint\b/i.test(prevBlock.text)
-  })
-
-  if (hasAuthenticatedUserRoute) {
-    state.results = routeBlocks.map((routeBlock) => {
-      const routeResult = _.merge({
-        method: routeBlock.method,
-        path: routeBlock.path,
-        enabledForApps: result.enabledForApps
-      }, result)
-
-      // <description>, <route block 1>, <description>, <route block 2>
-      // e.g. https://developer.github.com/v3/activity/starring/#list-repositories-being-starred
-      // => make description the name of the endpoints
-      const routeBlockIndex = state.blocks.indexOf(routeBlock)
-      const prevBlock = state.blocks[routeBlockIndex - 1]
-      routeResult.name = prevBlock.text.replace(/[^\w]*$/, '')
-      state.blocks.splice(routeBlockIndex - 1, 2)
-
-      return routeResult
-    })
-
-    return
-  }
-
-  if (hasExampleRoutes) {
-    routeBlocks.slice(1).forEach(block => {
-      block.type = 'description'
-      block.text = '```\n' + [block.method, block.path].join(' ') + '\n```'
-
-      delete block.method
-      delete block.path
-    })
-  }
-
-  if (hasStubbedRoute) {
-    result.method = routeBlocks[0].method
-    result.path = routeBlocks[0].path
-    const result2 = _.clone(result)
-    result2.name += ' (stubbed)'
-    result2.path = routeBlocks[1].path
-    state.results.push(result2)
-
-    const routeBlockIndex = state.blocks.indexOf(routeBlocks[1])
-    state.blocks.splice(routeBlockIndex - 1, 2)
-  }
-}
-
-/**
- * Find parameters: most endpoint have either no parameters or a single
- * parameters block, but some have two and more: https://github.com/gr2m/octokit-rest-routes/issues/4
- */
-function findParameters (state) {
-  const parametersBlocks = state.blocks.filter(block => block.type === 'parameters')
-  if (parametersBlocks.length === 0) {
-    state.results.forEach(result => {
-      result.params = []
-    })
-    return
-  }
-
-  const params = parametersBlocks[0].params
-  // remove paramters and heading above from blocks
-  state.blocks.splice(state.blocks.indexOf(parametersBlocks[0]) - 1, 2)
-
-  if (parametersBlocks.length === 1) {
-    state.results.forEach(result => {
-      result.params = params
-    })
-    return
-  }
-
-  parametersBlocks.slice(1).forEach(parametersBlock => {
-    const parametersBlockIndex = state.blocks.indexOf(parametersBlock)
-    const prevBlocks = state.blocks.slice(0, parametersBlockIndex - 1).reverse()
-    const alternativeHeaderBlock = prevBlocks.find(block => block.type === 'alternativeParametersHeader')
-    const optionalHeaderBlock = prevBlocks.find(block => block.type === 'optionalParametersHeader')
-
-    // "Alternative input" => make two endpoints
-    if (alternativeHeaderBlock) {
-      const result2 = _.clone(state.results[0])
-      state.results[0].params = params
-      result2.params = params
-      result2.name += ' (alternative)'
-      state.results.push(result2)
-
-      // remove parameters block, description & title above
-      state.blocks.splice(parametersBlockIndex - 1, 2)
-      return
-    }
-
-    // "Optional input" => add author & committer as optional parameters, turn table into description
-    if (optionalHeaderBlock) {
-      const optionalHeaderBlockIndex = state.blocks.indexOf(optionalHeaderBlock)
-      parametersBlock.type = 'description'
-
-      parametersBlock.text = markdownTable(
-        [['name', 'type', 'description']].concat(
-          parametersBlock.params.map(param => [param.name, param.type, param.description])
-        )
-      )
-
-      delete parametersBlock.params
-
-      // remove title and first paragraph from description above parameters table
-      state.blocks.splice(optionalHeaderBlockIndex, 2)
-
-      params.push({
-        name: 'committer',
-        type: 'object',
-        description: 'object containing information about the committer.'
-      })
-
-      params.push({
-        name: 'author',
-        type: 'object',
-        description: 'object containing information about the author.'
-      })
-
-      state.results.forEach(result => {
-        result.params = params
-      })
-      return
-    }
-
-    // "The xyz parameter takes the following keys"
-    // extend params with "<param>.<key>" | "<type>" | "<description>"
-    const prevBlock = state.blocks[parametersBlockIndex - 1]
-    const parameterName = prevBlock.text.match(/`([^`]+)`/).pop()
-    const parameterIndex = _.findIndex(params, (param) => param.name === parameterName)
-    params.splice(parameterIndex + 1, 0, ...parametersBlock.params.map(param => {
-      param.name = [parameterName, param.name].join('.')
-      return param
-    }))
-
-    // remove parameters block and description block above
-    state.blocks.splice(parametersBlockIndex - 1, 2)
-
-    state.results.forEach(result => {
-      result.params = params
-    })
-  })
-}
-
-function findDescription (state) {
-  const description = state.blocks
-    .filter(block => block.type === 'description')
-    .map(block => block.text)
-    .join('\n\n')
-
-  state.results.forEach(result => {
-    result.description = description
-  })
 }
