+++ lib/endpoint/tags/div.alert.note.js
@@ -2,7 +2,7 @@ const cheerio = require('cheerio')
 
 const turndown = require('../../turndown')
 
-const REGEX_PREVIEW_ACCEPT_HEADER = /\bapplication\/vnd\.github\.([^\s]+)-preview[^\s]+/
+const REGEX_PREVIEW_ACCEPT_HEADER = /\bapplication\/vnd\.github\.([^\s]+)-preview[^\s]*/
 const REGEX_PREVIEW_LABEL = /(\. | \n\s*)(.*)in the `Accept` header:[^]*/
 
 module.exports = {
