+++ lib/endpoint/tags/p.js
@@ -9,7 +9,8 @@ module.exports = {
   parse (el) {
     return {
       type: `description`,
-      text: turndown(cheerio.html(el))
+      text: turndown(cheerio.html(el)),
+      textOnly: cheerio(el).text()
     }
   }
 }
