+++ lib/endpoint/tags/table.js
@@ -5,7 +5,7 @@ const markdownTable = require('markdown-table')
 const turndown = require('../../turndown')
 
 const REQUIRED_REGEXP = /^\*\*Required\*\*\. /
-const CAN_BE_ONE_OF_REGEXP = /^Can be one of (`[a-z]+`)(, `[a-z]+`)*/
+const CAN_BE_ONE_OF_REGEXP = /Can be one of/
 const EITHER_REGEXP = /^Either (`[a-z]+`)/
 
 module.exports = {
@@ -68,7 +68,7 @@ function rowToParameter (i, el) {
 
         defaultValue = _defaultValue || (text.match(/\bDefault: (.*)$/) || []).pop()
 
-        // if description after the default value continous with a lowercase letter
+        // if description after the default value continues with a lowercase letter
         // it is like diretly related to it, e.g.
         //  > Default: true when environment is production and false otherwise.
         // https://developer.github.com/v3/repos/deployments/#create-a-deployment
@@ -90,8 +90,7 @@ function rowToParameter (i, el) {
 
       let enumValues
       if (CAN_BE_ONE_OF_REGEXP.test(description)) {
-        // enumValues = description.replace('Can be one of ', '').replace(/ or /g, '').split(/[, \.]/).filter((term) => term[0] === '`').map((term) => term.substring(1, term.length - 1))
-        enumValues = description.replace(/[,.]/g, '').split(' ').filter((term) => term[0] === '`' && term[term.length - 1] === '`').map((term) => term.substring(1, term.length - 1))
+        enumValues = _.uniq(description.match(/`([^`]+)`/g).map(s => s.replace(/`/g, '')))
       } else if (EITHER_REGEXP.test(description)) {
         // Include only words that begin and end with backticks (ignore all others)
         enumValues = description.replace(/[,.]/g, '').split(' ').filter((term) => term[0] === '`' && term[term.length - 1] === '`').map((term) => term.substring(1, term.length - 1))
